requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.24.1,<0.25
Pillow>=9.2.0,<9.3.0
uwsgi>=2.0.20<2.1
django-cors-headers>=3.13.0,<3.14
django-apscheduler>=0.6.2,<0.7